Modify build codes to add GFS de-GRIB feature
The default value for the -degrib_path option to `configure` is
based on the value of SSEC_ECCODES_LIB (an environment variable).

Also, add `configure` defaults for libhim_ifort and libhim_gcc.
Note that the default values for the -libhim_ifort and/or -libhim_gcc
options to `configure` are empty strings -- thus the appropriate one of
these options must be given a value by the user.