Fix array type passed to NetCDF OISST file reads
The three fields read from the NetCDF-format OISST file ('sst', 'err',
'ice') are all 2-byte integer ("short") fields in the files.  However,
in "main_src/oisst_analysis.f90", a 4-byte integer array (and thus a
4-byte integer NetCDF routine) is currently used to read each of those
fields.  This is not a good thing to do in general, causes a segfault
when using the Intel Fortran compiler, and may be silently causing
memory and/or other issues when using the gfortran compiler.  It is easy
to fix, however -- just declare a 2-byte integer array and use it for
these operations.