Fix array type in read_and_unscale_netcdf_*d
In the file 'cx_netcdf4_mod.f90', four subroutines
(read_and_unscale_netcdf_*d) are using the wrong array type to read in
data.  They state that they "Read and Unscale xD arrays Two-Byte Integers",
but they pass a 4-byte floating-point array to the NetCDF read routine.
Fixed by passing a 2-byte integer array to the NetCDF read routine.
This is not a good thing to do in general, causes a segfault
when using the Intel Fortran compiler, and may be silently causing
memory and/or other issues when using the gfortran compiler.