Modify NetCDF "string" global attribute read code
NetCDF "string" attributes are not reliably readable using the standard
Fortran NetCDF interface, and yet they are being used more frequently in
data files.  To fix this shortcoming, Yue Li added a C-language routine
that reads these NetCDF attribute values and a Fortran-2003 wrapper for
that routine.
For this commit, the Fortran-2003 wrapper was moved into the 'cx_netcdf4_mod'
module (where it is used).  The new build system was able to be quickly
adapted to compile the C-language NetCDF-dependent source code. Also,
this new feature currently only reads relevant NetCDF global attributes
(i.e., not variable attributes), and the names of the routines were
changed slightly to reflect this.