Removed the use of a compiler-specific intrinsic
The 'long()' intrinsic function is a gfortran-specific one (i.e., the
Intel Fortran compiler does not have it).  Furthermore, its function
is to typecast a < 4-byte integer to a 4-byte integer (the C 'long'
type) -- which may not be what was actually desired in this code, as it
seems like typecasting to an 8-byte integer would be more useful.

The Fortran-standard modification that was made uses the KINDs in the
'univ_kind_defs_mod' module (which is auto-created by the build process to
pick up compiler idiosyncracies), and assumes that 8-byte integer divsion is what was desired.  Other code could benefit from using that module as well.