Workaround for a subtle gfortran-related crash
For gfortran (at least v8.5.0), a pesky issue was found that led to
crashes during NOAA VIIRS granule processing, but not when using the
Intel compilers).  It was found that this could be averted if a separate
character string variable was used as input to the `h5readdataset` routine
(instead of performing an inline string concatenation when calling that
routine).  This workaround may also a modest number of CPU cycles.
Also, initialized all Fortran pointers upon declaration.