Deal better with out-of-bounds lunar phase angle
In routine `compute_lunar_reflectance`, the input variable
'lunar_phase_angle_topo' can be set to a missing-type value (-999.) and so
could not be found within the range of the 'phase_array' values -- this
led to trying to access an (improper) array index of zero.  The routine
now returns immediately after setting the output variable 'ref_chdnb_lunar'
to a canary value of -999.0 (consistent with the coding elsewhere in the
routine).