Change runtime example creation
Before, the runtime configuration file examples were tracked in run/.
Now they are tracked in src/files_for_run_time/ and freshly copied to run/
every time CLAVR-x is built.  This is so that if the user desires a
"clean/fresh" run/ directory, they can delete run/ and then rebuild CLAVR-x
to create a new one.
Also, prepare for adding example runscripts, which will be tracked and
copied similarly.