Change large array allocation (static -> dynamic)
More recent versions of the gfortran compiler emit the following warning for
large arrays that are statically-allocated: "Array is larger than limit set by
‘-fmax-stack-var-size=’, moved from stack to static storage. This makes the
procedure unsafe when called recursively, or concurrently from multiple threads.Consider using ‘-frecursive’, or increase the ‘-fmax-stack-var-size=’ limit,
or change the code to use an ALLOCATABLE array."

This commit changes the allocation of such arrays to be done
dynamically instead of statically.

For current typical CLAVR-x usage these types of warnings are probably
benign, but in the future they may not be if some sort of parallel
processing capability is introduced.