BUGFIX: 'land_fine_mode' set from empty array
An 'aerosol_land_map' array was statically declared (but only within the
scope of a single subroutine) with a shape of (season_index,nlon,nlat).
It was never assigned any values, but was then used to set
'this%land_fine_mode' using indices that did not match its declaration.
It is unclear what effect this had on CLAVR-x results.

This issue was fixed by using 'this%aerosol_land_map' (which has the
correct shape) instead of the local uninitialized array with the
conflicting shape.