Add build/process_template_file.py utility script
This is a Python script intended to simplify/automate the creation of
one or more files containing one or more text blocks that are sufficiently
similar ("templatable") -- all from two minimalistic input files:
1) a Python-syntax file that contains substitution keys/values that will be
iteratively applied to the templated text, and
2) a free-syntax file that contains templated text with substitution
tokens indicated by $token or ${token} ('$$' results in a literal '$'
character).
This code requires a single command-line argument, the shared prefix
(including path, if needed) of the two input files.